requirements.txt

flask
pycryptodome
pillow
